Allow custom external exclusions file for modsecurity rules
Since Home Assistant support an nearly unlimited assortment of integration and add-ons, it would not be possible to write an all-inclusive exclusions file for every possible false positive to that could be encountered.

Furthermore, it can be very difficult to reproduce false positives across systems. For example, if a false positive is encountered while using the Alexa integration, this could only be reproduced on another system that has the Alexa integration configured, which involves creating an Amazon developer account, creating a custom Alexa skill, writing Lamba functions, etc.

With this changes, users can write and maintain their own custom exclusions rules, allowing a path to fix their own false positives without rquiring a new release of the add-on.

Fixes #6